+++ Changelog.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 
+## [3.2.2b] - 2025-03-05
+
+### Fixed
+
+- [GroupManager] Progress data will now be visually split between NP and LOU progress.
+
+- [GroupManager] Ready check status will now be correctly updated on all group members.
+
+- [GroupManager] After clicking on a sort button the group manager will correctly update the visual data.
+
+- [ApplicationViewer] Pressing the refresh button will now scroll to the beginning of the list instead of erroring out.
+
+- LOU / NP progress in the [GroupManager] and [SearchPanel] progress panel will now be correctly attributed to the correct raid.
+
+
+
 ## [3.2.2a] - 2025-03-05
 
 ### Fixed